# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/approve": {
            "post": {
                "description": "approve progresss",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "upload"
                ],
                "summary": "approve a progress",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "req",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/controller.ApproveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/presign": {
            "get": {
                "description": "presign from oss",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "upload"
                ],
                "summary": "presign a data file",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "req",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/controller.PresignRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.ApproveRequest": {
            "type": "object",
            "required": [
                "approver"
            ],
            "properties": {
                "approver": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "opionion": {
                    "type": "string"
                },
                "trigger": {
                    "type": "string",
                    "default": "current"
                }
            }
        },
        "controller.PresignRequest": {
            "type": "object",
            "required": [
                "database",
                "filename",
                "filesize",
                "md5",
                "table"
            ],
            "properties": {
                "database": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "filesize": {
                    "type": "string"
                },
                "md5": {
                    "type": "string"
                },
                "table": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}